수정 삭제 권한 추가

+++ src/main/java/com/example/miniproject/service/ProductService.java
@@ -12,10 +12,14 @@ import com.example.miniproject.repository.MemberRepository;
 import com.example.miniproject.repository.ProductRepository;
 import com.example.miniproject.s3Service.S3Uploader;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -33,8 +37,8 @@ public class ProductService {
     private final MemberRepository memberRepository;
 
     @Transactional
-    public void createProduct(MultipartFile multipartFile, ProductRequestDto productRequestDto,
-                              String username) throws IOException {
+    public ResponseEntity<String> createProduct(MultipartFile multipartFile, ProductRequestDto productRequestDto,
+                                                String username) throws IOException {
         String imgUrl = s3Uploader.upload(multipartFile, "upload");
         Member member = memberRepository.findByUsername(username).orElseThrow();
 
@@ -50,6 +54,7 @@ public class ProductService {
                 .modifiedAt(LocalDateTime.now())
                 .build();
         productRepository.save(post);
+        return new ResponseEntity<>("글이 등록되었습니다.", HttpStatus.CREATED);
     }
 
     @Transactional(readOnly = true)
@@ -89,14 +94,39 @@ public class ProductService {
     }
 
     @Transactional
-    public void updateProduct(Long productId, ProductRequestDto productRequestDto) {
+    public ResponseEntity<String> updateProduct(Long productId, ProductRequestDto productRequestDto, HttpServletRequest request) {
+        RequestToken requestToken = new RequestToken(request); // servelet에서 토큰 가져오기
+
+        Member member = memberRepository.findByUsername(requestToken.getUsername().orElseThrow(
+                () -> new IllegalArgumentException("Can not find username"))).orElseThrow();
+
         Post post = productRepository.findById(productId)
                 .orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
+
+        if (!post.getNickname().equals(member.getNickname())) {
+            return new ResponseEntity<>("작성자만 수정할 수 있습니다", HttpStatus.UNAUTHORIZED);
+        }
         post.updatePost(productRequestDto);
+
+        return new ResponseEntity<>("글이 수정되었습니다.", HttpStatus.OK);
+
     }
 
     @Transactional
-    public void deleteProduct(Long productId) {
+    public ResponseEntity<String> deleteProduct(Long productId, HttpServletRequest request) {
+        RequestToken requestToken = new RequestToken(request); // servelet에서 토큰 가져오기
+
+        Member member = memberRepository.findByUsername(requestToken.getUsername().orElseThrow(
+                () -> new IllegalArgumentException("Can not find username"))).orElseThrow();
+
+        Post post = productRepository.findById(productId)
+                .orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
+
+        if (!post.getNickname().equals(member.getNickname())) {
+            return new ResponseEntity<>("작성자만 삭제할 수 있습니다", HttpStatus.UNAUTHORIZED);
+        }
         productRepository.deleteById(productId);
+
+        return new ResponseEntity<>("글이 삭제되었습니다.", HttpStatus.NO_CONTENT);
     }
 }
